add dummy models + reanem controller (remove survey prefix on filename not on class name)
