--- lib/app/modules/survey/controllers/survey_splash_controller.dart
@@ -1,20 +1,0 @@
-import 'package:get/get.dart';
-
-class SurveySplashController extends GetxController {
-  var loading = true.obs;
-  var loadingMessage = "Loading master data...".obs;
-
-  @override
-  void onInit() {
-    // TODO: implement onInit
-    super.onInit();
-    _synchronizeData();
-  }
-
-  void _synchronizeData() async {
-    await Future.delayed(Duration(seconds: 2));
-    loadingMessage.value = "Done!";
-    await Future.delayed(Duration(seconds: 1));
-    loading.value = false;
-  }
-}
